% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{as_latex}
\alias{as_latex}
\alias{as_latex.default}
\alias{as_latex.latex}
\alias{as_latex.NULL}
\title{Convert objects to LaTeX}
\usage{
as_latex(x, ...)

\method{as_latex}{default}(x, ...)

\method{as_latex}{latex}(x, ...)

\method{as_latex}{NULL}(x, ...)
}
\arguments{
\item{x}{The object to convert}

\item{...}{Other arguments used by methods}
}
\value{
An object of class \code{"tex"}.
}
\description{
This is the preferred method to convert objects to a
\code{latex} object. It is a generic function, so it can be defined for
different classes. The default is to convert an object to a character vector
and use \code{escape_latex} to escape special LaTeX symbols.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: This converts a character vector to a \code{LaTeX} object.
Unlike \code{\link{LaTeX}}, it can, and by default, escapes special LaTeX
characters.

\item \code{latex}: Return \code{x}, so as to not valid LaTeX text.

\item \code{NULL}: This returns ""
}}

\seealso{
\code{\link{LaTeX}} for a description of code \code{"latex"} objects.
}
