% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{as_tex}
\alias{as_tex}
\alias{as_tex.default}
\alias{as_tex.tex}
\alias{as_tex.NULL}
\title{Convert objects to LaTeX}
\usage{
as_tex(x, ...)

\method{as_tex}{default}(x, ...)

\method{as_tex}{tex}(x, ...)

\method{as_tex}{NULL}(x, ...)
}
\arguments{
\item{x}{The object to convert}

\item{...}{Other arguments used by methods}
}
\value{
An object of class \code{"tex"}.
}
\description{
This is the preferred method to convert objects to a
\code{tex} object. It is a generic function, so it can be defined for
different classes. The default is to convert an object to a character vector
and use \code{escape_latex} to escape special LaTeX symbols.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: This converts a character vector to a \code{tex} object.
Unlike \code{\link{tex}}, it can, and by default, escapes special LaTeX
characters.

\item \code{tex}: This simply returns \code{x}, so it will not escape already
escaped text.

\item \code{NULL}: This returns ""
}}

\seealso{
\code{\link{tex}} for a description of code \code{"tex"} objects.
}
