% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_case.R
\name{bundle_case}
\alias{bundle_case}
\title{Assign most common-case variants of case-insensitive clusters to all observations}
\usage{
bundle_case(x)
}
\arguments{
\item{x}{An atomic vector (for \code{enframe()}) or a data frame with one or two columns
(for \code{deframe()}).}
}
\value{
A logical value or vector
}
\description{
Where entities or themes are referenced in the data in various ways with respect to case, it is often desirable to group these together for analysis.  This function does so while preserving the most common case for prettier reading and to avoid the need for look-ups when reporting or visualising.
}
\examples{
bundle_case(c('london','London','London','Paris','PARIS','paris','Paris'))
}
\author{
Robin Edwards
}
