% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner.r
\name{cleaner}
\alias{cleaner}
\title{Clean textual data}
\usage{
cleaner(
  rawtext,
  stopwords = "italian",
  delete_url = TRUE,
  lowercase = TRUE,
  simbols = TRUE,
  numbers = TRUE,
  smallwords = TRUE,
  spaces = TRUE
)
}
\arguments{
\item{rawtext}{texts to clean}

\item{stopwords}{stopwords language. Default to "italian". It supports the following languages:
danish, dutch, english, finnish, french, german, hungarian, italian, norwegian, portuguese, russian, spanish, and swedish}

\item{delete_url}{delete URLs in the texts (default TRUE)}

\item{lowercase}{convert words to lower (default TRUE)}

\item{simbols}{keep just alphanumeric characters (default TRUE).}

\item{numbers}{remove numbers (default TRUE).}

\item{smallwords}{remove words composed of up to 2 characters (default TRUE).}

\item{spaces}{remove extra white spaces (default TRUE).}
}
\value{
a text vector with cleaned textual data
}
\description{
Given a text vector, clean text is returned
}
\examples{
\dontrun{
clean_text <- cleaner(rawtext)}

}
