% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterizer.R
\name{clusterizer}
\alias{clusterizer}
\title{Text clustering}
\usage{
clusterizer(
  df,
  docid_field = NULL,
  text_field = NULL,
  min_docfreq = 0.5,
  max_docfreq = 99,
  tfidf = TRUE,
  element = "documents",
  k = NULL,
  k.max = NULL,
  nstart = 25,
  method = "kmeans",
  hc_method = NULL,
  return_fit = FALSE
)
}
\arguments{
\item{df}{a data frame with at least a column with textual data and a column with documents' ID}

\item{docid_field}{name of the column (in quotation marks) containing the IDs of the documents (default NULL)}

\item{text_field}{name of the column (in quotation marks) containing textual data}

\item{min_docfreq}{minimum values of a feature's document frequency, below which features will be removed (default 0.5 percentile)}

\item{max_docfreq}{maximum values of a feature's document frequency, above which features will be removed (default 99 percentile)}

\item{tfidf}{term frequency inverse document frequency weighting (default TRUE)}

\item{element}{elements to cluster. Available options are "documents" and "features" (default "documents")}

\item{k}{desired number of clusters (default NULL).
If NULL, the silhouette method is used to estimate the appropriate number of clusters}

\item{k.max}{max number of cluster if k is not specified (default NULL)}

\item{nstart}{number of initial configurations (default 25)}

\item{method}{clustering method. Available options are "kmeans" and "hclust" (default "kmeans)}

\item{hc_method}{the agglomeration method to be used in case of "hclust" method.
This should be one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA),
"median" (= WPGMC) or "centroid" (= UPGMC). See \link[stats]{hclust}}

\item{return_fit}{return the fitted cluster_model in the main environment}
}
\value{
an vector of cluster IDs. Silhouette information with clusters' size and average silhouette width are printed in console
}
\description{
Given a data frame with texts, documents (or features) clustering is returned
}
\details{
the function is substantially a wrapper of functions available in \href{https://quanteda.io/}{quanteda} and  \href{https://rpkgs.datanovia.com/factoextra/index.html}{factoextra}.
Please refer to the available documentations of \link[quanteda]{textstat_simil} and \link[factoextra]{eclust}.
The fitted cluster_model can be used to create kmeans plots with \link[factoextra]{fviz_cluster} or dendrogram (hclust) wiht \link[factoextra]{fviz_dend}
}
\examples{
\dontrun{
df$cluster <- clusterizer(df, docid_field = "documents", text_field = "texts", k = 10)}

}
