% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.table.R
\name{corr.table}
\alias{corr.table}
\title{Correlation Table}
\usage{
corr.table(
  x,
  type = c("pearson", "spearman"),
  tex = TRUE,
  title = NULL,
  labs = NULL,
  digits = 3,
  ...
)
}
\arguments{
\item{x}{a matrix or a data.frame with the variables for which the pairwise correlations should
be calculated.}

\item{type}{type of correlation coefficient that will be calculated.
One of "pearson" (default) and "spearman".}

\item{tex}{a logical value that indicates whether the results should be displayed as a LaTeX
table or not. Default is TRUE.}

\item{title}{a character variable that is used as the caption of the LaTeX table if tex is TRUE.
If title is NULL (default) then the title is "Correlation Table".}

\item{labs}{a character vector that is used for the rownames and the colnames of the table.
If labs is NULL (default) then the colnames of x are used for the rownames and colnames.}

\item{digits}{number of digits that are displayed for relative frequencies. Default is 3.}

\item{...}{further arguments to be passed to other functions.}
}
\value{
\code{corr.table} uses \code{stargazer} to return LaTeX code for a table.
}
\description{
corr.table creates a lower triangle correlation table with stars that indicate whether a
correlation coefficient is significantly different from zero.
}
