% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctable.R
\name{ctable}
\alias{ctable}
\title{Contingency Table}
\usage{
ctable(
  x,
  y,
  margin = 2,
  tex = FALSE,
  title = NULL,
  rowlabs = NULL,
  collabs = NULL,
  digits = 1,
  ...
)
}
\arguments{
\item{x}{A vector that represents the variable displayed in rows.}

\item{y}{A vector that represents the variable displayed in columns.}

\item{margin}{Whether the relative frequencies should be computed rowwise
or columnwise}

\item{tex}{Whether the results should be displayed as a LaTeX table.}

\item{title}{Title for the LaTeX table.}

\item{rowlabs}{A vector for the names of the rows in LaTeX table. 
If rowlabs is NULL then the rownames that are shown in the output 
of \code{summary} or \code{Anova} are use.}

\item{collabs}{Names for the columns.}

\item{digits}{Number of digits that are displayed in the LaTeX tables.}

\item{...}{Further optional arguments.}
}
\value{
\code{ctable} uses \code{stargazer} to return LaTeX code for a table.
}
\description{
\code{ctable} creates a contingency table with absolute and relative
frequencies.
}
