% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textools.R
\name{escape_latex}
\alias{escape_latex}
\title{Escape LaTeX special characters}
\usage{
escape_latex(x, url = TRUE, ellipses = TRUE)
}
\arguments{
\item{x}{Character vector}

\item{url}{If \code{TRUE}, escape URLs by enclosing them in \verb{\url} macros.}

\item{ellipses}{If \code{TRUE}, replace \verb{...}.}
}
\value{
A character vector with all LaTeX special characters escaped.
}
\description{
Escape special LaTeX special characters in text.
}
\details{
The following characters are escaped as follows:
\tabular{ll}{
Orginal       \tab  Escaped     \cr
\verb{\{}     \tab  \verb{\\\{} \cr
\verb{\}}     \tab  \verb{\\\}} \cr
\verb{#}      \tab  \verb{\\#}  \cr
\verb{$}      \tab  \verb{\\$}  \cr
\verb{&}      \tab  \verb{\\&}  \cr
\verb{_}      \tab  \verb{\\_}  \cr
\verb{\%}     \tab  \verb{\\\%} \cr
\verb{\\}     \tab  \verb{\\textbackslash{}} \cr
\verb{~}      \tab  \verb{\\textasciitilde{}} \cr
\verb{^}      \tab  \verb{\\textasciicircum{}} \cr
\verb{...}    \tab  \verb{\\dots}
}
}

