% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr2latex.R
\name{expr2latex}
\alias{expr2latex}
\alias{expr2latex_}
\title{Convert Expression to LaTeX}
\usage{
expr2latex(...)

expr2latex_(expr)
}
\arguments{
\item{...}{Unquoted expressions}

\item{expr}{Language objects, atomic values, or lists. If \code{expr} is a list,
it will recursively call \code{expr2latex_} on each element.}
}
\value{
A character vector with the LaTeX version of the
   expression.
}
\description{
Convert an R expression to a LaTeX mathematical equation.
}
\details{
Most of the code from this function adapted from the example in
\href{http://adv-r.had.co.nz/dsl.html}{R for Data Science},
"Domain Specific Languages".

The \code{symbol} function uses the \verb{\Pisymbol} command
from the \verb{pifont} package. You will need to load it in
in any code that uses these commands.
}
\author{
Hadley Wickham
}
\references{
http://adv-r.had.co.nz/dsl.html
}
\seealso{
\code{\link{plotmath}}
}

