% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IN.R
\name{\%IN\%}
\alias{\%IN\%}
\title{String in text (vectorised)}
\usage{
q \%IN\% str
}
\value{
A logical value or vector
}
\description{
The text-in operator
}
\examples{
x = rownames(mtcars)
x['merc' \%IN\% x]
x[ gsub(' .*', '', x) \%IN\% 'We have Mazdas and Fiats in stock' ]
x[purrr::map_lgl(x, ~ any(c('merc','fiat') \%IN\% .x))]
}
\author{
Robin Edwards
}
