% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{is_tex_command}
\alias{is_tex_command}
\title{Check that string is a valid TeX command name}
\usage{
is_tex_command(x)
}
\arguments{
\item{x}{character vector}
}
\value{
A logical vector indicating which elements elements of the vector
  contain valid TeX command names.

Without resorting to changing the category codes of characters, valid
LaTeX commands can only contain letters.

Assertation to check for a valid LaTeX macro (command) names
}
\description{
Check that string is a valid TeX command name
}
