% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{is_tex_name}
\alias{is_tex_name}
\alias{make_tex_name}
\title{Check or make valid TeX command name}
\usage{
is_tex_name(x, allow_at = FALSE)

make_tex_name(x, allow_at = FALSE)
}
\arguments{
\item{x}{character vector}

\item{allow_at}{flag. Allow the character vector to include the \code{@}
character in addition to upper- and lower-cased letters.}
}
\value{
\describe{
\item{\code{is_tex_name}}{A logical vector with length \code{length(x)}
indicating which elements elements of the vector contain valid TeX
command names}
\item{}{A character vector with length \code{length(x)} with syntactically
  valid TeX command names.}
}
}
\description{
Valid LaTeX command names can only contain letters \code{[A-Za-z]}.
The function \code{is_tex_name} checks whether strings are valid LaTeX
command names. The function \code{make_tex_name} converts a character vector
to valid LaTeX command names by dropping all invalid characters.
}
