% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex}
\alias{latex}
\alias{latex.default}
\title{Convert R object to LaTeX}
\usage{
latex(x, ...)

\method{latex}{default}(x, escape = TRUE, ...)
}
\arguments{
\item{x}{An R object to be converted to LaTeX.}

\item{...}{Arguments passed to methods.}

\item{escape}{Escape LaTeX using the function \code{\link{escape_latex}}.}
}
\value{
An object of class \code{c("latex", "character")}.
}
\description{
Convert R object to LaTeX
}

