% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{latex_macros}
\alias{latex_macros}
\alias{[[<-.latex_macros}
\title{Create and Write LaTeX Macro Lists}
\usage{
latex_macros(...)

\method{[[}{latex_macros}(x, i) <- value
}
\value{
An object of class \code{latex_macros} which extends
 \code{environment}.
}
\description{
Create and Write LaTeX Macro Lists
}
\section{Methods (by generic)}{
\itemize{
\item \code{[[<-}: This checks for the validity `i` before
adding it.
}}

