% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_macros.R
\docType{data}
\name{LaTeXMacroList}
\alias{LaTeXMacroList}
\title{Create and Write LaTeX Macro Lists}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LaTeXMacroList
}
\description{
This is an R6 object which simplifies using \code{\link{list_to_macros}}
for writing R objects to LaTeX macros. These objects have simple methods
for adding and dropping elements in the list, and to write macros.
}
\keyword{datasets}
