% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.R
\name{lca}
\alias{lca}
\title{Lexical Correspondence Analysis}
\usage{
lca(
  df,
  docid_field = NULL,
  text_field = NULL,
  min_docfreq = 0.5,
  max_docfreq = 99
)
}
\arguments{
\item{df}{a dataframe with at least a column with textual data and a column with documents' ID}

\item{docid_field}{name of the column (in quotation marks) containing the IDs of the documents (default NULL)}

\item{text_field}{name of the column (in quotation marks) containing textual data}

\item{min_docfreq}{minimum values of a feature's document frequency, below which features will be removed (default 0.5 percentile)}

\item{max_docfreq}{maximum values of a feature's document frequency, above which features will be removed (default 99 percentile)}
}
\value{
a correspondence analysis model
}
\description{
correspondence analysis of textual data
}
\details{
the function is substantially a wrapper of functions available in \href{https://quanteda.io/}{quanteda} and \href{https://rpkgs.datanovia.com/factoextra/index.html}{factoextra}.
More specifically it leverages the correspondence analysis function \link[FactoMineR]{CA}.
The fitted ca_model can be used to further explore the model and create plots with \link[factoextra]{fviz_ca_biplot}
}
\examples{
\dontrun{
dataframe$cluster <- lca(df, docid_field = "documents", text_field = "texts")}

}
