% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_macros.R
\name{list_to_macros}
\alias{list_to_macros}
\title{Convert a list to LaTeX macros}
\usage{
list_to_macros(x, prefix = "", collapse = "\\n", ...)
}
\arguments{
\item{x}{A list or vector}

\item{prefix}{Prefix added to names}

\item{collapse}{If not \code{NULL}, then macros are collapsed to
a single string with \code{collapse} as a seperator.}

\item{...}{Arguments passed to \code{\link{texnewcmd}}.}
}
\description{
For a list or vector. For elements with non-missing names,
the elements are converted to a string and t
}

