% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textools.R
\name{ltxcmd}
\alias{as.character.latex_command}
\alias{ltxcmd}
\alias{macro}
\title{LaTeX Macros}
\usage{
ltxcmd(command, args = NULL, ..., optargs = NULL)

\method{as.character}{latex_command}(x, ..., trailing = TRUE)

macro(...)
}
\arguments{
\item{command}{character Name of the LaTeX command}

\item{args}{character vector.}

\item{...}{Additional arguments in methods}

\item{optargs}{character vector containing optional arguments which go in
square brackets. Any named elments are converted to key=value
pairs as is common in many macros. If \code{length(args) > 1},
then arguments are comma separated.}

\item{x}{In methods, the \code{macro} object.}

\item{trailing}{If \code{TRUE}, then trailing brackets will be
added to the macro even if there are no arguments.}
}
\value{
An object of class \code{"latex_command"}. This is a \code{list}
  with elements \code{cmd}, \code{args}, and \code{optargs}.
}
\description{
This returns text wrapped in a LaTeX macro:
\verb{\\cmd[optional]\{arg[1]\}\{arg[2]\}}.
}

