% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textools.R
\name{ltxenv}
\alias{begin}
\alias{ltxenv}
\title{LaTeX Environment}
\usage{
ltxenv(name, content, args = NULL, optargs = NULL)

begin(...)
}
\arguments{
\item{name}{Environment name}

\item{content}{Text to go in the environment.}

\item{args}{Environment arguments}

\item{optargs}{Optional environment arguments (argument inside brackets)}

\item{...}{Other arguments needed for method definitions}
}
\value{
An object of class \code{"latex_environment"}, which is a list with
  elements \code{name}, \code{content}, \code{args}, and \code{optargs}.
}
\description{
This an object represeting a LaTeX environment and which can be rendered
to a LaTeX environment.
}

