% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinert.R
\name{reinert}
\alias{reinert}
\title{Reinert clustering (Rainette)}
\usage{
reinert(
  df,
  docid_field = NULL,
  text_field = NULL,
  segment_size = 40,
  k = NULL,
  min_uc_size = 10,
  min_split_members = 5,
  cc_test = 0.3,
  tsj = 3
)
}
\arguments{
\item{df}{a data frame with at least a column with textual data and a column with documents' ID}

\item{docid_field}{name of the column (in quotation marks) containing the IDs of the documents (default NULL)}

\item{text_field}{name of the column (in quotation marks) containing textual data}

\item{min_uc_size}{minimum number of forms by document (default 10).}

\item{min_split_members}{don't try to split groups with fewer members (default 5).}

\item{cc_test}{contingency coefficient value for feature selection (default 0.3).}

\item{tsj}{minimum frequency value for feature selection (default 3).}

\item{k.max}{desired maximum number of clusters (default NULL).}
}
\value{
a list containing the result of clustering and the original document term matrix, for subsequent analysis
through \link[rainette]{rainette_explor} or \link[rainette]{rainette_plot}
}
\description{
Given a data frame with texts, this function perform Reinert's clustering
}
\details{
the function is simply a wrapper of functions available in the package \href{https://juba.github.io/rainette/}{rainette}.
Please refer to the available documentations of \link[rainette]{rainette}
}
\examples{
\dontrun{
reinert_clustering <- reinert(df, docid_field = "URL", text_field = "Message", k = 10)
rainette::rainette_explor(reinert_clustering[[1]], reinert_clustering[[2]])}

}
