% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_detect_multi.R
\name{str_detect_multi}
\alias{str_detect_multi}
\title{Detect the presence or absence of multiple patterns in a string/vector}
\usage{
str_detect_multi(strings, patterns, ignore_case = TRUE, regex = TRUE)
}
\arguments{
\item{ignore_case}{Should case differences be ignored in the match?}
}
\value{
A logical value or vector
}
\description{
A wrapper for stringr::str_detect to enable vectorised search against multiple terms, with regex and case_ignore applied by default
}
\examples{
x = rownames(mtcars)
x[str_detect_multi(x, c('merc','honda'))]

require(dplyr)
starwars \%>\% filter(str_detect_multi(name, c('sky','darth')))
}
\author{
Robin Edwards
}
