% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_groups.R
\name{summary_groups}
\alias{summary_groups}
\title{LaTeX tables for summaries by groups}
\usage{
summary_groups(
  x,
  groups,
  sum.all = TRUE,
  rowlabs = NULL,
  tex = FALSE,
  title = NULL,
  digits = 2,
  ...
)
}
\arguments{
\item{x}{The variable for which a summary is computed.}

\item{groups}{A group variable.}

\item{sum.all}{Whether a summary over all groups should be added.}

\item{rowlabs}{A vector for the names of the rows in LaTeX table. 
If rowlabs is NULL then the rownames that are shown in the output 
of \code{summary} or \code{Anova} are use.}

\item{tex}{Whether the output should be TeX code.}

\item{title}{Title that is used as the caption of the LaTeX table. 
If \code{title = NULL} then the caption is "Descriptive Summary".}

\item{digits}{Number of digits that are displayed in the LaTeX tables.}

\item{...}{Further optional arguments.}
}
\value{
\code{summary_groupss} uses \code{stargazer} to return 
  LaTeX code for a table.
}
\description{
\code{summary_groupss} generates a LaTeX table with the descriptive
summary of a continuous variable by groups.
}
