% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_uni.R
\name{table_uni}
\alias{table_uni}
\title{LaTeX table for univariate frequency tables}
\usage{
table_uni(x, tex = FALSE, title = NULL, rowlabs = NULL, digits = 1, ...)
}
\arguments{
\item{x}{A categorial vector.}

\item{tex}{Whether the results should be displayed as a LaTeX table.}

\item{title}{Title for the LaTeX table.}

\item{rowlabs}{A vector for the names of the rows in LaTeX table. 
If rowlabs is NULL then the rownames that are shown in the output 
of \code{summary} or \code{Anova} are use.}

\item{digits}{Number of digits that are displayed in the LaTeX tables.}

\item{...}{Further optional arguments.}
}
\value{
\code{table_uni} uses \code{stargazer} to return LaTeX code for a table.
}
\description{
\code{table_uni} creates LaTeX tables for univariate frequency
tables.
}
