% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texcmd.R
\name{texarg}
\alias{texarg}
\alias{texargs}
\alias{texopt}
\title{LaTeX Arguments}
\usage{
texarg(x, delim = "{", kv = TRUE)

texopt(...)

texargs(args = list(), ..., optargs = character(), kv = TRUE)
}
\arguments{
\item{x}{An object}

\item{delim}{A character vector of length one or two representing
the delimiters used to enclose the argument when printing.
If the element is either \code{"\{"} or \code{"["}, then the
matching closing indicator will be used, otherwise the element will be used
as both the opening and closing delimiters. If two elements are specified,
then the two elements are used as a delimiter. If \code{NULL} or \code{TRUE},
then curly brackets (braces) are used for required arguments. If \code{FALSE}, then
square brackets are used for optional arguments.}

\item{kv}{The argument in \code{x} is concatenated into a list of options and
key-value pairs using \code{\link{texkv}}.}

\item{...}{Additional arguments}

\item{args}{Either a single argument or a list of arguments}

\item{optargs}{Optional arguments for a LaTeX command. These are enclosed in
in square brackets and placed before any of the arguments in \code{args}
or \code{...}.}

\item{kv}{Arguments are converted to a comma-seperated option
and key-value list.}
}
\value{
An object of class \code{"texarg"}. This is character vector
  with a single element, with attributes \code{"open"} and \code{"close"}
  for the opening and closing delimeters for the argument.
}
\description{
Functions and classes to represent and format arguments and
lists of arguments for a for a LaTeX command. This has easy options for
to wrap text for LaTeX required (\verb{\{\}}) and optional (\verb{[]}) arguments,
but is written generally since LaTeX is so flexible that macros can do anything.
}
\section{Functions}{
\itemize{
\item \code{texopt}: Utility function for the common case of optional
LaTeX arguments. It is \code{texarg} with \code{delim = "["}.

\item \code{texargs}: \code{texargs} creates a list of \code{texarg} objects.
This is largely used for easy of printing.
}}

