% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texcmd.R
\name{texargs}
\alias{texargs}
\title{LaTeX arguments to a macro}
\usage{
texargs(x, escape = TRUE, nargs = TRUE, ...)
}
\arguments{
\item{x}{A list or character object convertible by \code{latex}}

\item{escape}{If \code{TRUE}, then \code{x} will be LaTeX escaped.}

\item{nargs}{If \code{x} is \code{NULL} or has a length shorter than
\code{nargs}, extra empty arguments are added. If \code{nargs} is less
than the length of \code{x}, then it is ignored. This is to make it easy
to generate empty argument lists list \verb{\{\}\{\}\{\}}.}

\item{...}{Arguments passed to \code{latex}}
}
\description{
This is just a unnamed character vector with a special
\code{latex} method. It also specifies that the elements
within it are already valid LaTeX.
}

