% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texcmd.R
\name{texcmd}
\alias{texcmd}
\title{LaTeX commands (macros)}
\usage{
texcmd(name, ..., .kv = TRUE)
}
\arguments{
\item{name}{Name of the LaTeX command}

\item{...}{Command arguments. These are passed to \code{\link{texargs}}.
Elements should be either a character vector, \code{"tex"} object,
\code{"texarg"} object, or a list with arguments for the \code{texargs}
function. By default arguments are assumed to be mandatory. Use the
\code{texopt} function to create optional args.}
}
\description{
This returns text wrapped in a LaTeX macro:
\verb{\\cmd[optional]\{arg[1]\}\{arg[2]\}}.
}
\details{
The function \code{texcmd} returns a list with the elements
in the LaTeX macro, while the function \code{texcmd_} returns
the rendered LaTeX string as a character vector;
executing \code{texcmd(x)} is equivalent to \code{format(texcmd_(x))}.
}

