% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texcmd.R
\name{texenv}
\alias{texenv}
\title{LaTeX Environments}
\usage{
texenv(name, body = "", args = list(), optargs = list(), ...)
}
\arguments{
\item{name}{Environment name}

\item{body}{Text content of the macro}

\item{args}{Arguments to the environment. See \code{\link{texarg}}.}

\item{optargs}{Optional arguments for the environment. These are arguments
what will be enclosed in square brackets \code{[]}, and preceed arguments in
\code{args}.}

\item{...}{Arguments passed to \code{\link{as.tex}} to coerce
\code{body} into a \code{tex} object.}
}
\value{
An object of class \code{"texenv"}, which is a list with
 named elements: \code{name} (\code{"character"}),
 \code{body} (\code{"character")}), \code{"args"} (\code{"texargs"}).
}
\description{
Create an object representing text enclosed in a LaTeX environment.
}

