% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{texkv}
\alias{texkv}
\title{Key-value argument list}
\usage{
texkv(x, ..., .escape = FALSE)
}
\arguments{
\item{x}{A character vector or list. Elements are converted to tex objects
via \code{\link{as.tex}}.}

\item{...}{Additional possibly named arguments}

\item{.escape}{Whether to LaTeX escape the string values.}
}
\value{
A \code{\link{tex}} object.
}
\description{
Create an argument list of comma separated options and key-value pairs
as is often seen in LaTeX commands.
For example, "arg1, arg2=value2, arg3, ...".
}

