% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texmod.aov.R
\name{texmod.aov}
\alias{texmod.aov}
\title{LaTeX tables for aov models}
\usage{
\method{texmod}{aov}(
  mod,
  results = c("summary", "Anova"),
  ssq = TRUE,
  meansq = TRUE,
  df = TRUE,
  teststatistic = FALSE,
  pval = TRUE,
  title = NULL,
  n_title = TRUE,
  rowlabs = NULL,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{aov}.}

\item{results}{Whether the displayed results are from \code{summary}
or from \code{Anova} from the package car.}

\item{ssq}{Whether the sum of squares should be displayed 
if \code{results = "Anova"}.}

\item{meansq}{Whether the mean squares should be displayed if
\code{results = "summary"}.}

\item{df}{Whether the degrees of freedom should be displayed if 
\code{results = "Anova"}.}

\item{teststatistic}{Whether the test statistic should be displayed.}

\item{pval}{Whether the p-values should be displayed.}

\item{title}{Title that is used as the caption of the LaTeX table. 
If \code{title = NULL} then the caption is ”Analysis of Variance”.}

\item{n_title}{Wheter the sample size that was used to compute the model should
be added to the caption of the table.}

\item{rowlabs}{A vector for the names of the rows in LaTeX table. 
If rowlabs is NULL then the rownames that are shown in the output 
of \code{summary} or \code{Anova} are use.}

\item{digits}{Number of digits that are displayed in the LaTeX tables.}

\item{...}{Further optional arguments.}
}
\description{
texmod method for models of class \code{aov}.
}
\examples{
iris.aov <- aov(Sepal.Length ~ Sepal.Width + Petal.Width + Species, data = iris)
texmod(iris.aov,
 title = "Analysis of Variance for Sepal Length",
 rowlabs = c("Sepal Width", "Petal Width", "Species")
)
}
