% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texmod.coxph.R
\name{texmod.coxph}
\alias{texmod.coxph}
\title{LaTeX tables for coxph models}
\usage{
\method{texmod}{coxph}(
  mod,
  results = c("summary", "Anova"),
  hr = TRUE,
  loghr = FALSE,
  ci = TRUE,
  ci_level = 0.95,
  se_loghr = FALSE,
  teststatistic = FALSE,
  df = TRUE,
  test = c("LR", "Wald"),
  pval = TRUE,
  title = NULL,
  n_title = TRUE,
  rowlabs = NULL,
  addref = TRUE,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{coxph}.}

\item{results}{Whether the displayed results are from \code{summary}
or from \code{Anova} from the package car.}

\item{hr}{Whether the hazard ratios should be displayed if 
\code{results = "summary"}.}

\item{loghr}{Whether the log hazard ratios should be displayed if 
\code{results = "summary"}.}

\item{ci}{Whether confidence intervals for the hazard ratios should be 
displayed if \code{results = "summary"}.}

\item{ci_level}{Level for the confidence intervals.}

\item{se_loghr}{Whether the standard errors of the log harzard ratios 
should be displayed if \code{results = "summary"}.}

\item{teststatistic}{Whether the test statistic should be displayed.}

\item{df}{Whether the degrees of freedom should be displayed if 
\code{results = "Anova"}.}

\item{test}{Which test should be calculated, if \code{results = "Anova"}.
Either \code{"LR"} or \code{"Wald"}.}

\item{pval}{Whether the p-values should be displayed.}

\item{title}{Title that is used as the caption of the LaTeX table. 
If title is NULL then the caption is ”Cox Regression” if 
\code{results = "summary"} and ”Cox Analysis of Variance” if 
\code{results = "Anova"}.}

\item{n_title}{Wheter the sample size that was used to compute the model should
be added to the caption of the table.}

\item{rowlabs}{A vector for the names of the rows in LaTeX table. 
If rowlabs is NULL then the rownames that are shown in the output 
of \code{summary} or \code{Anova} are use.}

\item{addref}{Whether a row for the reference category of categorial variables 
should be displayed if \code{results = "summary"}.}

\item{digits}{Number of digits that are displayed in the LaTeX tables.}

\item{...}{Further optional arguments.}
}
\description{
texmod method for models of class \code{coxph}.
}
\examples{
library(survival)
lung.coxph <- coxph(Surv(time, status) ~ age + factor(sex) + meal.cal + ph.karno,
  data = lung)
texmod(lung.coxph,
  title = "Cox Regression for Overall Survival",
  rowlabs = c("Age", "Sex (male)", "Sex (female)", "Meal calories",
    "Karnofsky score (physician)")
 )
}
