% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texmod.crr.R
\name{texmod.crr}
\alias{texmod.crr}
\title{LaTeX tables for crr models}
\usage{
\method{texmod}{crr}(
  mod,
  hr = TRUE,
  loghr = FALSE,
  ci = TRUE,
  ci_level = 0.95,
  se_loghr = FALSE,
  teststatistic = FALSE,
  pval = TRUE,
  title = NULL,
  rowlabs = NULL,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{crr}.}

\item{hr}{Whether the hazard ratios should be displayed if 
\code{results = "summary"}.}

\item{loghr}{Whether the log hazard ratios should be displayed if 
\code{results = "summary"}.}

\item{ci}{Whether confidence intervals for the hazard ratios should be 
displayed if \code{results = "summary"}.}

\item{ci_level}{Level for the confidence intervals.}

\item{se_loghr}{Whether the standard errors of the log harzard ratios 
should be displayed if \code{results = "summary"}.}

\item{teststatistic}{Whether the test statistic should be displayed.}

\item{pval}{Whether the p-values should be displayed.}

\item{title}{Title that is used as the caption of the LaTeX table. 
If title is NULL then the caption is ”Competing Risks Regression".}

\item{rowlabs}{A vector for the names of the rows in LaTeX table. 
If rowlabs is NULL then the rownames that are shown in the output 
of \code{summary} or \code{Anova} are use.}

\item{digits}{Number of digits that are displayed in the LaTeX tables.}

\item{...}{Further optional arguments.}
}
\description{
texmod method for models of class \code{crr}.
}
