% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texmod.glm.R
\name{texmod.glm}
\alias{texmod.glm}
\title{LaTeX tables for glm models}
\usage{
\method{texmod}{glm}(
  mod,
  results = c("summary", "Anova"),
  or = TRUE,
  logor = FALSE,
  ci = TRUE,
  ci_level = 0.95,
  se.logor = FALSE,
  vcov = NULL,
  teststatistic = FALSE,
  df = TRUE,
  test = c("LR", "Wald", "F"),
  sq = TRUE,
  pval = TRUE,
  intercept = FALSE,
  title = NULL,
  n_title = TRUE,
  rowlabs = NULL,
  addref = TRUE,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{glm}.}

\item{results}{Whether the displayed results are from \code{summary}
or from \code{Anova} from the package car.}

\item{logor}{Whether the log odds ratios should be displayed if
\code{results = "summary"}.}

\item{ci}{Whether confidence intervals for the odds ratios should be 
displayed if \code{results = "summary"}.}

\item{ci_level}{Level for the confidence intervals.}

\item{vcov}{If \code{NULL} then the covariance matrix computed
from \code{lm} will be used. Alternatively, one of the covariance matrix
functions from the the \code{sandwich} package can be passed.}

\item{teststatistic}{Whether the test statistic should be displayed.}

\item{df}{Whether the degrees of freedom should be displayed if 
\code{results = "Anova"}.}

\item{test}{Which test should be calculated, if \code{results = "Anova"}.
Either \code{"LR"}, \code{"Wald"} or \code{"F"}.}

\item{pval}{Whether the p-values should be displayed.}

\item{intercept}{Whether a row for the intercept should be 
displayed if \code{results = "summary"}.}

\item{title}{Title that is used as the caption of the LaTeX table. 
If \code{title = NULL} then the caption is ”Logistic Regression” if 
\code{results = "summary"} and ”Binomial Analysis of Variance” if 
\code{results = "Anova"}.}

\item{n_title}{Wheter the sample size that was used to compute the model should
be added to the caption of the table.}

\item{rowlabs}{A vector for the names of the rows in LaTeX table. 
If rowlabs is NULL then the rownames that are shown in the output 
of \code{summary} or \code{Anova} are use.}

\item{addref}{Whether a row for the reference category of categorial variables 
should be displayed if \code{results = "summary"}.}

\item{digits}{Number of digits that are displayed in the LaTeX tables.}

\item{...}{Further optional arguments.}

\item{hr}{Whether the odds ratios should be displayed if
\code{results = "summary"}.}

\item{se_logor}{Whether the standard errors of the log odds ratios
should be displayed if \code{results = "summary"}.}

\item{ssq}{Whether the sum of squares should be displayed 
if \code{results = "Anova"} and \code{test = "F"}.}
}
\description{
texmod method for models of class \code{glm}.
}
\details{
Models of class \code{glm} are currently only supported for 
  logistic regression model.
}
\examples{
iris$Sepal.Length <- ifelse(iris$Sepal.Length < median(iris$Sepal.Length), 0, 1)
iris.glm <- glm(Sepal.Length ~ Sepal.Width + Petal.Width + Species, 
  data = iris,
  family = binomial
  )
texmod(iris.glm,
  title = "Logistic Regression for median(Sepal Length)",
  rowlabs = c("Sepal Width", "Petal Width", "Species (versicolor)",
  "Species (setosa)", "Species (virginica)")
)
}
