% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texmod.glmerMod.R
\name{texmod.glmerMod}
\alias{texmod.glmerMod}
\title{LaTeX tables for glmerMod models}
\usage{
\method{texmod}{glmerMod}(
  mod,
  results = c("summary", "Anova"),
  or = TRUE,
  logor = FALSE,
  ci = TRUE,
  ci_level = 0.95,
  se_logor = FALSE,
  teststatistic = FALSE,
  df = TRUE,
  pval = TRUE,
  intercept = FALSE,
  title = NULL,
  n_title = TRUE,
  rowlabs = NULL,
  addref = TRUE,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{glmerMod}.}

\item{results}{Whether the displayed results are from \code{summary}
or from \code{Anova} from the package car.}

\item{or}{Whether the odds ratios should be displayed if
\code{results = "summary"}.}

\item{logor}{Whether the log odds ratios should be displayed if
\code{results = "summary"}.}

\item{ci}{Whether confidence intervals for the odds ratios should be 
displayed if \code{results = "summary"}.}

\item{ci_level}{Level for the confidence intervals.}

\item{se_logor}{Whether the standard errors of the log odds ratios
should be displayed if \code{results = "summary"}.}

\item{teststatistic}{Whether the test statistic should be displayed.}

\item{df}{Whether the degrees of freedom should be displayed if 
\code{results = "Anova"}.}

\item{pval}{Whether the p-values should be displayed.}

\item{intercept}{Whether a row for the intercept should be 
displayed if \code{results = "summary"}.}

\item{title}{Title that is used as the caption of the LaTeX table. 
If \code{title = NULL} then the caption is ”Logistic Mixed Model Regression” 
if \code{results = "summary"} and ”Binomial Mixed Model ANOVA” if 
\code{results = "Anova"}.}

\item{n_title}{Wheter the sample size that was used to compute the model should
be added to the caption of the table.}

\item{rowlabs}{A vector for the names of the rows in LaTeX table. 
If rowlabs is NULL then the rownames that are shown in the output 
of \code{summary} or \code{Anova} are use.}

\item{addref}{Whether a row for the reference category of categorial variables 
should be displayed if \code{results = "summary"}.}

\item{digits}{Number of digits that are displayed in the LaTeX tables.}

\item{...}{Further optional arguments.}
}
\value{
\code{texph} uses \code{stargazer} to return LaTeX code for a table.
}
\description{
texmod method for models of class \code{glmerMod}.
}
\details{
Models of class \code{glmerMod} are currently only supported for 
  logistic mixed model regressions.
}
\examples{
bdf <- nlme::bdf
bdf$IQ.verb <- ifelse(bdf$IQ.verb < median(bdf$IQ.verb), 0, 1)
bdf.glmerMod <- lme4::glmer(IQ.verb ~ sex + aritPOST + Minority +
   (1|schoolNR), data = bdf, family = "binomial")
texmod(bdf.glmerMod,
  method = "Wald",
  title = "Logistic Mixed Model Regression for Verbal IQ",
  rowlabs = c("Sex (female)", "Sex (male)", "Arit (post)",
    "Minority (no)", "Minority (yes)")
)
}
