% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texmod.lm.R
\name{texmod.lm}
\alias{texmod.lm}
\title{LaTeX tables for lm models}
\usage{
\method{texmod}{lm}(
  mod,
  results = c("summary", "Anova"),
  estimate = TRUE,
  ci = TRUE,
  ci_level = 0.95,
  se = FALSE,
  vcov = NULL,
  teststatistic = FALSE,
  ssq = TRUE,
  df = TRUE,
  pval = TRUE,
  intercept = FALSE,
  title = NULL,
  n_title = TRUE,
  rowlabs = NULL,
  addref = TRUE,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{lm}.}

\item{results}{Whether the displayed results are from \code{summary}
or from \code{Anova} from the package car.}

\item{estimate}{Whether the estimates (corresponding to the intercept and the 
slopes) should be displayed if \code{results = "summary"}.}

\item{ci}{Whether confidence intervals for the estimates should be displayed 
if \code{results = "summary"}.}

\item{ci_level}{Level for the confidence intervals.}

\item{se}{Whether the standard errors of the estimates should be
displayed if \code{results = "summary"}.}

\item{vcov}{If \code{NULL} then the covariance matrix computed
from \code{lm} will be used. Alternatively, one of the covariance matrix
functions from the the \code{sandwich} package can be passed.}

\item{teststatistic}{Whether the test statistic should be displayed.}

\item{ssq}{Whether the sum of squares should be displayed 
if \code{results = "Anova"}.}

\item{df}{Whether the degrees of freedom should be displayed if 
\code{results = "Anova"}.}

\item{pval}{Whether the p-values should be displayed.}

\item{intercept}{Whether a row for the intercept should be 
displayed if \code{results = "summary"}.}

\item{title}{Title that is used as the caption of the LaTeX table. 
If \code{title = NULL} then the caption is ”Linear Regression” if 
\code{results = "summary"} and ”Analysis of Variance” if 
\code{results = "Anova"}.}

\item{n_title}{Wheter the sample size that was used to compute the model should
be added to the caption of the table.}

\item{rowlabs}{A vector for the names of the rows in LaTeX table. 
If rowlabs is NULL then the rownames that are shown in the output 
of \code{summary} or \code{Anova} are use.}

\item{addref}{Whether a row for the reference category of categorial variables 
should be displayed if \code{results = "summary"}.}

\item{digits}{Number of digits that are displayed in the LaTeX tables.}

\item{...}{Further optional arguments.}
}
\value{
\code{texph} uses \code{stargazer} to return LaTeX code for a table.
}
\description{
texmod method for models of class \code{lm}.
}
\examples{
iris.lm <- lm(Sepal.Length ~ Sepal.Width + Petal.Width + Species, data = iris)
texmod(iris.lm,
  title = "Linear Model for Sepal Length",
  rowlabs = c("Sepal Width", "Petal Width", "Species (versicolor)",
    "Species (setosa)", "Species (virginica)")
)
}
