% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texmod.lme.R
\name{texmod.lme}
\alias{texmod.lme}
\title{LaTeX tables for lme models}
\usage{
\method{texmod}{lme}(
  mod,
  results = c("summary", "Anova"),
  estimate = TRUE,
  ci = TRUE,
  ci_level = 0.95,
  se = FALSE,
  teststatistic = FALSE,
  df = TRUE,
  pval = TRUE,
  intercept = FALSE,
  title = NULL,
  n_title = TRUE,
  rowlabs = NULL,
  addref = TRUE,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{lme}.}

\item{results}{Whether the displayed results are from \code{summary}
or from \code{Anova} from the package car.}

\item{estimate}{Whether the estimates (corresponding to the fixed 
intercept and the the fixedslopes) should be displayed if 
\code{results = "summary"}.}

\item{ci}{Whether confidence intervals for the estimates should be displayed 
if \code{results = "summary"}.}

\item{ci_level}{Level for the confidence intervals.}

\item{se}{Whether the standard errors of the estimates should be
displayed if \code{results = "summary"}.}

\item{teststatistic}{Whether the test statistic should be displayed.}

\item{df}{Whether the degrees of freedom should be displayed if 
\code{results = "Anova"}.}

\item{pval}{Whether the p-values should be displayed.}

\item{intercept}{Whether a row for the intercept should be 
displayed if \code{results = "summary"}.}

\item{title}{Title that is used as the caption of the LaTeX table. 
If \code{title = NULL} then the caption is ”Mixed Model Regression” if 
\code{results = "summary"} and ”Mixed Model Analysis of Variance” 
if \code{results = "Anova"}.}

\item{n_title}{Wheter the sample size that was used to compute the model should
be added to the caption of the table.}

\item{rowlabs}{A vector for the names of the rows in LaTeX table. 
If rowlabs is NULL then the rownames that are shown in the output 
of \code{summary} or \code{Anova} are use.}

\item{addref}{Whether a row for the reference category of categorial variables 
should be displayed if \code{results = "summary"}.}

\item{digits}{Number of digits that are displayed in the LaTeX tables.}

\item{...}{Further optional arguments.}
}
\description{
texmod method for models of class \code{lme}.
}
\examples{
library(nlme)
bdf.lme <- nlme::lme(IQ.verb ~ sex + aritPOST + denomina + Minority,
  random = ~1|schoolNR,
  data = nlme::bdf)
texmod(bdf.lme,
  title = "Mixed Model Regression for Verbal IQ",
  rowlabs = c("Sex (female)", "Sex (male)", "Arit (post)", 
    "School (public)", "School (protestant)", "School (catholic)",
    "School (private)", "Minority (no)", "Minority (yes)")
)
}
