% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texmod.lmerModLmerTest.R
\name{texmod.lmerModLmerTest}
\alias{texmod.lmerModLmerTest}
\title{LaTeX tables for lmerModLmerTest models}
\usage{
\method{texmod}{lmerModLmerTest}(
  mod,
  results = c("summary", "Anova"),
  estimate = TRUE,
  ci = TRUE,
  ci.level = 0.95,
  se = FALSE,
  teststatistic = FALSE,
  df = TRUE,
  df.res = TRUE,
  test = c("Chisq", "F"),
  pval = TRUE,
  intercept = FALSE,
  title = NULL,
  n_title = TRUE,
  rowlabs = NULL,
  addref = TRUE,
  digits = 3,
  ...
)
}
\arguments{
\item{results}{Whether the displayed results are from \code{summary}
or from \code{Anova} from the package car.}

\item{estimate}{Whether the estimates (corresponding to the fixed 
intercept and the the fixedslopes) should be displayed if 
\code{results = "summary"}.}

\item{ci}{Whether confidence intervals for the estimates should be displayed 
if \code{results = "summary"}.}

\item{se}{Whether the standard errors of the estimates should be
displayed if \code{results = "summary"}.}

\item{teststatistic}{Whether the test statistic should be displayed.}

\item{df}{Whether the degrees of freedom should be displayed if 
\code{results = "Anova"} or when \code{results = "summary"} and
\code{pval = TRUE}.}

\item{test}{Which test should be calculated, if \code{results = "Anova"}.
Either \code{"Chisq"} or \code{"F"}.}

\item{pval}{Whether the p-values should be displayed.}

\item{intercept}{Whether a row for the intercept should be 
displayed if \code{results = "summary"}.}

\item{title}{Title that is used as the caption of the LaTeX table. 
If \code{title = NULL} then the caption is ”Mixed Model Regression” if 
\code{results = "summary"} and ”Mixed Model Analysis of Variance” 
if \code{results = "Anova"}.}

\item{n_title}{Wheter the sample size that was used to compute the model should
be added to the caption of the table.}

\item{rowlabs}{A vector for the names of the rows in LaTeX table. 
If rowlabs is NULL then the rownames that are shown in the output 
of \code{summary} or \code{Anova} are use.}

\item{addref}{Whether a row for the reference category of categorial variables 
should be displayed if \code{results = "summary"}.}

\item{digits}{Number of digits that are displayed in the LaTeX tables.}

\item{...}{Further optional arguments.}

\item{A}{model of class \code{lmerModLmerTest}.}

\item{ci_level}{Level for the confidence intervals.}

\item{df_res}{Whether degrees of freedom residuals should be displayed if 
\code{results = "Anova"} and \code{test = "F"}.}
}
\value{
\code{texph} uses \code{stargazer} to return LaTeX code for a table.
}
\description{
texmod method for models of class \code{lmerModLmerTest}.
}
\details{
If \code{pval = TRUE} then the package lmerTest is called.
}
\examples{
bdf.lmerMod <- lmerTest::lmer(IQ.verb ~ sex + aritPOST + denomina + 
  Minority + (1|schoolNR), data = nlme::bdf)
texmod(bdf.lmerMod,
  title = "Mixed Model Regression for Verbal IQ",
  rowlabs = c("Sex (female)", "Sex (male)", "Arit (post)", 
    "School (public)", "School (protestant)", "School (catholic)", 
    "School (private)", "Minority (no)", "Minority (yes)")
)
}
