% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texmod.R
\name{texmod}
\alias{texmod}
\title{LaTeX tables for regression models}
\usage{
texmod(mod, ...)
}
\arguments{
\item{mod}{A model object for which a LaTeX table should created.}

\item{...}{Further optional arguments.}
}
\description{
texmod is a generic function that creates LaTeX tables
for various regression models.
}
