% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texnewcmd.R
\name{texnewcmd}
\alias{newcommand}
\alias{providecommand}
\alias{renewcommand}
\alias{texnewcmd}
\title{LaTeX newcommand macro}
\usage{
texnewcmd(name, definition, nargs = 0, default = NULL,
  command = c("providecommand", "renewcommand", "newcommand"),
  starred = FALSE)

providecommand(...)

newcommand(...)

renewcommand(...)
}
\arguments{
\item{name}{Command name}

\item{definition}{Definition of the command}

\item{nargs}{Number of arguments the command takes}

\item{default}{Default value for an optional argument}

\item{command}{Command to use to create the LaTeX command.}

\item{starred}{If \code{TRUE}, use the starred versions of the
commands, e.g. \verb{\\\\providecommand*}.}

\item{...}{Arguments passed to \code{texnewcmd} in
\code{newcommand} and \code{renewcommand}.}
}
\value{
An object of class \code{"texnewcmd"},
}
\description{
LaTeX newcommand macro
}

