% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texargs.R
\name{texopts}
\alias{texopts}
\title{LaTeX Option List}
\usage{
texopts(x, ..., escape = TRUE)
}
\arguments{
\item{x}{A list or character vector.}

\item{...}{Arguments passed to \code{\link{latex}} when
converting elements in \code{x}.}

\item{escape}{If \code{TRUE}, then the values of x are
latex escaped.}
}
\value{
An object of class \code{"texopts"}.
}
\description{
Create an object representing LaTeX the values in LaTeX
optional or default arguments. Arguments appearing in
optional or default arguments often take the form of
a comma-separated list of either named options, `key=value`,
or simply flags, `value`.
}
\details{
The primary purpose of this class is to take a vector like
\code{c(opta = "val1", "optb", ...)}
with optionally named elements and produce a character
string like \code{"[opta=val1, optb, ...]"}.
}

