% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texph.coxph.R
\name{texph.coxph}
\alias{texph.coxph}
\title{Post-hoc LaTeX tables for coxph models}
\usage{
\method{texph}{coxph}(
  mod,
  variable,
  pairwise = TRUE,
  estimate = TRUE,
  se = FALSE,
  teststatistic = FALSE,
  pval = TRUE,
  ci = TRUE,
  ci_level = 0.95,
  title = NULL,
  varlab = NULL,
  rowlabs = NULL,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{coxph}.}

\item{variable}{}

\item{pairwise}{}

\item{estimate}{}

\item{se}{}

\item{teststatistic}{}

\item{pval}{}

\item{ci}{}

\item{ci_level}{}

\item{title}{}

\item{varlab}{}

\item{rowlabs}{}

\item{digits}{}

\item{...}{}
}
\description{
texph method for models of class \code{coxph}.
}
\details{
If \code{ci = TRUE} the displayed confidence intervals 
  are Wald-type confidence intervals.
}
\examples{
lung <- survival::lung
lung$age.group <- with(lung, ifelse(age < 55, "< 55",
  ifelse(age <65, "55 to 65", "> 65")))
lung.coxph <- survival::coxph(survival::Surv(time, status) ~ age.group + 
  factor(sex) + meal.cal,
  data = lung)
texph(lung.coxph,
  variable = "age.group",
  title = "Pairwise Hazard Ratios for Overall Survival"
)

texph(lung.coxph,
  pairwise = FALSE,
  variable = "age.group",
  varlab = "Age Group",
  title = "EM Hazards for Overall Survival"
)
}
