% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texph.glmerMod.R
\name{texph.glmerMod}
\alias{texph.glmerMod}
\title{Post-hoc LaTeX tables for glmerMod models}
\usage{
\method{texph}{glmerMod}(
  mod,
  variable,
  pairwise = TRUE,
  estimate = TRUE,
  se = FALSE,
  teststatistic = FALSE,
  pval = TRUE,
  ci = TRUE,
  ci_level = 0.95,
  title = NULL,
  varlab = NULL,
  rowlabs = NULL,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{glmerMod}.}

\item{variable}{}

\item{pairwise}{}

\item{estimate}{}

\item{se}{}

\item{teststatistic}{}

\item{pval}{}

\item{ci}{}

\item{ci_level}{}

\item{title}{}

\item{varlab}{}

\item{rowlabs}{}

\item{digits}{}

\item{...}{}
}
\value{

}
\description{
texph method for models of class \code{glmerMod}.
}
