% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texph.lm.R
\name{texph.lm}
\alias{texph.lm}
\title{Post-hoc LaTeX tables for lm models}
\usage{
\method{texph}{lm}(
  mod,
  variable,
  pairwise = TRUE,
  estimate = TRUE,
  se = FALSE,
  df = TRUE,
  teststatistic = FALSE,
  pval = TRUE,
  ci = TRUE,
  ci_level = 0.95,
  title = NULL,
  varlab = NULL,
  rowlabs = NULL,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{lm}.}

\item{variable}{The variable for which the pairwise comparisons or 
the estimated marginal means should be computed. Has to be a 
character.}

\item{pairwise}{Whether pairwise comparisons between all levels
of the variable or the estimated marginal means for each level 
of the variable should be displayed.}

\item{estimate}{Whether the estimate (corresponding to the estimated mean 
difference between the levels if \code{pairwise = TRUE} and the estimated 
marginal means if \code{pairwise = FALSE}) should be displayed.}

\item{se}{Whether the standard errors of the estimates should be
displayed.}

\item{df}{Whether the degrees of freedom should be displayed.}

\item{teststatistic}{Whether the test statistic should be displayed if
\code{pairwise = TRUE}.}

\item{pval}{Whether the p-values should be displayed if 
\code{pairwise = TRUE}. Note that the p-values are adjusted for
multiple testing. The default adjustment method is Tukey’s method.}

\item{ci}{Whether confidence intervals for the estimates should be 
displayed. If \code{pairwise = TRUE} the confidence intervals are
simultaneous and the default adjustment method is Tukey’s method. 
If \code{pairwise = FALSE} the confidence intervals are not 
simultaneous.}

\item{ci_level}{Level for the confidence intervals.}

\item{title}{Title that is used as the caption of the LaTeX table. 
If title is NULL then the caption is ”Pairwise Comparisons” if 
\code{pairwise = TRUE} and ”EM Means” if \code{pairwise = FALSE}.}

\item{varlab}{The column label for the levels of \code{variable} if 
\code{pairwise = FALSE}. If \code{varlab = NULL} then the name of 
\code{variable} is used.}

\item{rowlabs}{A vector for the names of the rows in LaTeX table. 
If rowlabs is NULL then the rownames that are shown in the output 
of \code{summary} or \code{Anova} are use.}

\item{digits}{Number of digits that are displayed in the LaTeX tables.}

\item{...}{Further optional arguments.}
}
\value{
\code{texph} uses \code{stargazer} to return LaTeX code for a table.
}
\description{
texph method for models of class \code{lm}.
}
\details{
The method for the multiple comparison adjustments for 
  both the p-values and the confidence intervals (only for pairwise 
  comparisons) can be changed by passing the \code{adjust} argument of 
  \code{emmeans} function through the \code{...} argument. The same
  adjustment method will be used for both p-values and confidence intervals. 
  The default method is Tukey’s method. Other possibilities are \code{"scheffe"}, 
  \code{"sidak"}, \code{"bonferroni"}, \code{"dunnettx"}, \code{"mvt"}, and 
  \code{"none"}. Please refer to the reference manual of the library emmeans for 
  details. Note that some other adjustment methods are available in emmeans for 
  p-values but not for confidence intervals. If these methods are used in
  \code{texph} the p-values will be adjusted by the correct method but the 
  simultaneous confidence intervals will be calculated using Bonferroni’s method 
  without producing a warning.
}
\examples{
iris.lm <- lm(Sepal.Length ~ Sepal.Width + Petal.Width + Species, 
  data = iris)
texph(iris.lm,
  variable = "Species",
  title = "Pairwise Comparisons for Sepal Length"
)

texph(iris.lm,
  pairwise = FALSE,
  variable = "Species",
  title = "EM Means for Sepal Length"
)
}
