% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texph.lmerModLmerTest.R
\name{texph.lmerModLmerTest}
\alias{texph.lmerModLmerTest}
\title{Post-hoc LaTeX tables for lmerModLmerTest models}
\usage{
\method{texph}{lmerModLmerTest}(
  mod,
  variable,
  pairwise = TRUE,
  estimate = TRUE,
  se = FALSE,
  df = TRUE,
  teststatistic = FALSE,
  pval = TRUE,
  ci = TRUE,
  ci_level = 0.95,
  title = NULL,
  varlab = NULL,
  rowlabs = NULL,
  digits = 3,
  ...
)
}
\arguments{
\item{mod}{A model of class \code{lmerModLmerTest}.}

\item{variable}{The variable for which the pairwise comparisons or 
the estimated marginal means should be computed. Has to be a 
character.}

\item{pairwise}{Whether pairwise comparisons between all levels
of the variable or the estimated marginal means for each level 
of the variable should be displayed.}

\item{estimate}{Whether the estimate (corresponding to the estimated mean 
difference between the levels if \code{pairwise = TRUE} and the estimated 
marginal means if \code{pairwise = FALSE}) should be displayed.}

\item{se}{Whether the standard errors of the estimates should be
displayed.}

\item{df}{Whether the degrees of freedom should be displayed.}

\item{teststatistic}{Whether the test statistic should be displayed if
\code{pairwise = TRUE}.}

\item{pval}{Whether the p-values should be displayed if 
\code{pairwise = TRUE}. Note that the p-values are adjusted for
multiple testing. The default adjustment method is Tukey’s method.}

\item{ci}{Whether confidence intervals for the estimates should be 
displayed. If \code{pairwise = TRUE} the confidence intervals are
simultaneous and the default adjustment method is Tukey’s method. 
If \code{pairwise = FALSE} the confidence intervals are not 
simultaneous.}

\item{ci_level}{Level for the confidence intervals.}

\item{title}{Title that is used as the caption of the LaTeX table. 
If title is NULL then the caption is ”Pairwise Comparisons” if 
\code{pairwise = TRUE} and ”EM Means” if \code{pairwise = FALSE}.}

\item{varlab}{The column label for the levels of \code{variable} if 
\code{pairwise = FALSE}. If \code{varlab = NULL} then the name of 
\code{variable} is used.}

\item{rowlabs}{A vector for the names of the rows in LaTeX table. 
If rowlabs is NULL then the rownames that are shown in the output 
of \code{summary} or \code{Anova} are use.}

\item{digits}{Number of digits that are displayed in the LaTeX tables.}

\item{...}{Further optional arguments.}
}
\value{
\code{texph} uses \code{stargazer} to return LaTeX code for a table.
}
\description{
texph method for models of class \code{lmerModLmerTest}.
}
\examples{
bdf.lmerMod <- lmerTest::lmer(IQ.verb ~ sex + aritPOST + denomina + 
  Minority + (1|schoolNR),
  data = nlme::bdf)
texph(bdf.lmerMod,
  variable = "denomina",
  rowlabs = c("Public - Protestant", "Public - Catholic", "Public - Private",
    "Protestant - Catholic", "Protestant - Private", "Catholic - Private"),
  title = "Pairwise Comparisons for Verbal IQ"
)

texph(bdf.lmerMod,
  variable = "denomina",
  pairwise = FALSE,
  varlab = "Denomina",
  title = "EM Means for Verbal IQ"
)
}
