% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texph.R
\name{texph}
\alias{texph}
\title{LaTeX tables for post-hoc comparisons}
\usage{
texph(mod, ...)
}
\arguments{
\item{mod}{A model object for which a LaTeX table should created.}

\item{...}{Further optional arguments.}
}
\description{
texph is a generic functions that creates LaTeX tables for post-hoc
comparisons and estimated marginal means of various regression models.
}
