% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withlatex.R
\name{with_latex}
\alias{with_latex}
\title{Evaluate an expression using \code{latex}}
\usage{
with_latex(code)
}
\arguments{
\item{code}{Code to evaluate with the \code{\link{latex}} environment.}
}
\description{
Evaluate an expression using \code{latex}
}

