% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_macros.R
\name{write_latex_commands}
\alias{write_latex_commands}
\alias{write_latex_commands.default}
\title{Convert a list to LaTeX macros}
\usage{
write_latex_commands(x, ...)

\method{write_latex_commands}{default}(x, prefix = "", ...)
}
\arguments{
\item{x}{A list or vector}

\item{...}{Arguments passed to \code{\link{as.tex}}.}

\item{prefix}{Prefix added to names}
}
\description{
For a list or vector. For elements with non-missing names,
the elements are converted to a string and t
}
