% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_case.R
\name{bundle_case}
\alias{bundle_case}
\title{Assign most common-case variants of case-insensitive clusters to all observations}
\usage{
bundle_case(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector of same length as input, with case variants changed to the mode
}
\description{
Where names/entities/themes/tags/etc are referenced with multiple variations of upper/lower/title/mixed case, it is often desirable to group these case-insensitively for analysis.  This function does so while preserving the most common case for prettier reading and to avoid the need for subsequent look-ups.
}
\examples{
bundle_case(c('london','London','London','Paris','PARIS','paris','Paris'))
}
\author{
Robin Edwards
}
