% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IN.R
\name{\%IN\%}
\alias{\%IN\%}
\title{String in text (vectorised)}
\usage{
q \%IN\% str
}
\arguments{
\item{q}{Regex pattern or vector of patterns to match against str}

\item{str}{Character value or vector to match q against}
}
\value{
A logical value or vector
}
\description{
A Python-style String in Text operator that ignores case.
Shorthand for str_detect(str, regex(q, ignore_case = TRUE)).
Either q or str must be a single value. The return is a logical value or vector if vectorised input is provided.
}
\examples{
'cat' \%IN\% "The Cat in a Hat"
c('cat', 'fox') \%IN\% "The Cat in a Hat"
'cat' \%IN\% c('The Cat in a Hat', 'Fox in Socks')

require(dplyr)
d = starwars \%>\% select(name) \%>\% as.data.frame()
d \%>\% filter('sky' \%IN\% name)
d \%>\% filter('[0-9]' \%IN\% name) # regex
}
\author{
Robin Edwards
}
