% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withlatex.R
\docType{data}
\name{latex}
\alias{latex}
\title{LaTeX Builder Functions}
\format{An object of class \code{list} of length 20.}
\usage{
latex
}
\description{
The \code{latex} environment contains some convenience functions
for creating LaTeX.
}
\keyword{datasets}

