% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex}
\alias{latex}
\alias{latex.default}
\title{Convert R object to LaTeX text}
\usage{
latex(x, ...)

\method{latex}{default}(x, ..., escape = TRUE)
}
\arguments{
\item{x}{An R object to be converted to LaTeX.}

\item{...}{Arguments passed to methods.}

\item{escape}{Escape LaTeX using the function \code{\link{escape_latex}}.}
}
\value{
An object of class \code{c("latex", "character")}. The
  \code{"latex"} class is primarily used by other functions in this
  package to identify text that is already LaTeX, so it knows not to
  escape LaTeX special characters.
}
\description{
Marks the given text as (La)TeX, which means functions will
know not to perform escaping on it.
}
\examples{
latex("Some non-escaped LaTeX code with special symbols like # and $.")
latex("Already \\\\textit{formatted} \\\\LaTeX text.", escape = FALSE)

}
\seealso{
\code{\link[utils]{toLatex}}
}

