% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{latex_macros}
\alias{latex_macros}
\title{Create and Write LaTeX Macro Lists}
\usage{
latex_macros(...)
}
\arguments{
\item{...}{For \code{latex_macros}, named arguments in which the argument
name is the name of LaTeX command the expression is converted to a string
using \code{\link{as_latex}}.  The dots are evaluated using [explicit splicing][dots_list].}
}
\value{
An object of class \code{latex_macros} which extends
 \code{environment}.
}
\description{
Create and Write LaTeX Macro Lists
}
\examples{
macros <- latex_macros(a = "1", b = "foo")
macros
# Assign new macros
macros$c <- 2 + 2
macros$d <- paste(letters, collapse = " ")
macros
# By default, LaTeX special characters are escaped
macros$e <- "$ 2 ^ 3 $"
macros$e
# use tex() to avoid escaping
macros$e <- tex("$ 2 ^ 3 $")
macros$e
# format() returns a string with LaTeX commands
format(macros)
}
