% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{latex_macros}
\alias{latex_macros}
\title{Create and Write LaTeX Macro Lists}
\usage{
latex_macros(...)
}
\arguments{
\item{...}{For \code{latex_macros}, named arguments in which the argument
name is the name of LaTeX command the expression is converted to a string
using \code{\link{as_tex}}.  The dots are evaluated using [explicit splicing][dots_list].}
}
\value{
An object of class \code{latex_macros} which extends
 \code{environment}.
}
\description{
Create and Write LaTeX Macro Lists
}
