% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{tex}
\alias{tex}
\title{Convert R object to LaTeX text}
\usage{
tex(x)
}
\arguments{
\item{x}{An R object to be converted to LaTeX.}

\item{...}{Arguments passed to methods.}
}
\value{
An object of class \code{c("latex", "character")}. The
  \code{"tex"} class is primarily used by other functions in this
  package to identify text that is already LaTeX, so it knows not to
  escape LaTeX special characters.
}
\description{
Marks the given text as (La)TeX, which means functions will
know not to perform escaping on it.
}
\examples{
tex("Already \\\\textit{formatted} \\\\LaTeX text.")
}
\seealso{
\code{\link[utils]{toLatex}}
}

