% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texargs.R
\name{texargs}
\alias{texargs}
\title{LaTeX Command Arguments}
\usage{
texargs(x, nargs = NULL, escape = TRUE, ...)
}
\arguments{
\item{x}{A list or character object. All elements in \code{x} are
passed to \code{tex} to convert to valid LaTeX.}

\item{nargs}{If \code{x} is \code{NULL} or has a length shorter than
\code{nargs}, extra empty arguments are added. If \code{nargs} is less
than the length of \code{x}, then it is ignored. This is to make it easy
to generate empty argument lists list \verb{\{\}\{\}\{\}}.}

\item{escape}{If \code{TRUE}, then \code{x} will be LaTeX escaped.}

\item{...}{Arguments passed to \code{tex}
when converting elements of \code{x}.}
}
\description{
The function \code{textargs} produces an object that represents
LaTeX argument lists. The primary purpose of this argument is
to easily print to the argument list to the proper LaTeX form
and to initialize empty arguments.
}
\examples{
# three arguments
texargs(c("a", "b", "c"))
# three arguments, with the last two empty
texargs(c("a"), nargs = 2)
# three empty arguments
texargs(nargs = 3)
}

