% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texcmd.R
\name{texmacro}
\alias{texcmd}
\alias{texmacro}
\title{LaTeX commands (macros)}
\usage{
texmacro(name, trailing = FALSE, starred = FALSE)

texcmd(name, ...)
}
\arguments{
\item{name}{Name of the LaTeX command}

\item{trailing}{Include trailing curly brackets}

\item{starred}{Make the macro starred}

\item{...}{Command arguments. These are passed to \code{\link{texargs}}.
Elements should be either a character vector, \code{"tex"} object,
\code{"texarg"} object, or a list with arguments for the \code{texargs}
function. By default arguments are assumed to be mandatory. Use the
\code{texopt} function to create optional args.}
}
\description{
This returns text wrapped in a LaTeX macro:
\verb{\\cmd[optional]\{arg[1]\}\{arg[2]\}}.
}
\details{
The function \code{texcmd} returns a list with the elements
in the LaTeX macro, while the function \code{texcmd_} returns
the rendered LaTeX string as a character vector;
executing \code{texcmd(x)} is equivalent to \code{format(texcmd_(x))}.
}
\section{Functions}{
\itemize{
\item \code{texmacro}: texmacro returns the string for a LaTeX symbol,
i.e. a command without arguments.
}}

