% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texenv.R
\name{texenv}
\alias{texenv}
\alias{texenv_}
\title{LaTeX Environment}
\usage{
texenv(name, content, args = NULL, opts = NULL)

texenv_(name, content, args = NULL, opts = NULL)
}
\arguments{
\item{name}{Environment name}

\item{content}{Text to go in the environment.}

\item{args}{Environment arguments}

\item{opts}{Optional environment arguments (argument inside brackets)}

\item{...}{Other arguments needed for method definitions}
}
\value{
An object of class \code{"texenv"}, which is a list with
  elements \code{name}, \code{content}, \code{args}, and \code{optargs}.
}
\description{
This an object represeting a LaTeX environment and which can be rendered
to a LaTeX environment.
}
\examples{
texenv("figure",
       texcmd("includegraphics", "figure.png"),
       optargs=c("h"))
texenv("center", "Centered text")
texenv("enumerate",
       latex(paste(c("\\\\item First item", "\\\\item Second item"),
                     collapse = "\\n"),
             escape = FALSE))
}

