% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texnewcmd.R
\name{texnewcmd}
\alias{texnewcmd}
\alias{texnewcmd_}
\title{LaTeX newcommand macro}
\usage{
texnewcmd(name, definition, nargs = 0, default = NULL,
  command = c("providecommand", "renewcommand", "newcommand"),
  starred = FALSE, escape = FALSE)

texnewcmd_(name, definition, nargs = 0, default = NULL,
  command = c("providecommand", "renewcommand", "newcommand"),
  starred = FALSE, escape = FALSE)
}
\arguments{
\item{name}{Command name}

\item{definition}{Definition of the command}

\item{nargs}{Number of arguments the command takes}

\item{default}{Default value for an optional argument}

\item{command}{Command to use to create the LaTeX command.}

\item{starred}{If \code{TRUE}, use the starred versions of the
commands, e.g. \verb{\\\\providecommand*}.}
}
\value{
The function \code{texnewcmd_} returns an object of class \code{"texnewcmd"},
   while is a list with elements \code{name}, \code{definition}, \code{nargs},
   \code{default}, \code{command}, and \code{starred}.
   The function \code{texnewcmd} returns an object of class \code{"tex"} with
   the LaTeX string.
}
\description{
LaTeX newcommand macro
}
\examples{
texnewcmd("absval",
          texcmd("ensuremath", tex("\\\\lvert#1\\\\rvert", FALSE)),
          nargs = 1)
}

