% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texnewenv.R
\name{texnewenv}
\alias{newenvironment}
\alias{renewenvironment}
\alias{texnewenv}
\title{LaTeX New Environment}
\usage{
texnewenv(name, begin_def = character(), end_def = character(), nargs = 0,
  default = NULL, command = c("newenvironment", "renewenvironment"),
  starred = FALSE)

newenvironment(...)

renewenvironment(...)
}
\arguments{
\item{name}{LaTeX environment name}

\item{begin_def}{Beginning definition of the environment}

\item{end_def}{End definition of the environment}

\item{nargs}{The number of arguments}

\item{default}{A default value for the first argument.}

\item{command}{The LaTeX command to use for defining an environment.}

\item{starred}{If \code{TRUE}, the use the starred versions of \code{command};
e.g. \verb{\\newenvironment*} and \verb{\\renewenvironment}.}

\item{...}{Arguments passed to \code{texnewenv} in  \code{newenvironment}
and \code{rewnewenvironment}.}
}
\value{
An object of class \code{latex_newenvironment}.
  This is a list with elements \code{name}, \code{begin_def},
  \code{nargs}, \code{default}, \code{command}, \code{starred}.
}
\description{
Create an object representating a \verb{\\newenvironment}
command in LaTeX.
}
\details{
When rendered, the object produces the following
LaTeX command
\verb{\\newenvironment{name}[nargs]default]{begin_def}{end_def}}.
}

