% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texcmd.R
\name{texopts}
\alias{texopts}
\title{LaTeX Option List}
\usage{
texopts(x, ..., escape = TRUE)
}
\arguments{
\item{x}{An list or character vector.}

\item{...}{Arguments passed to \code{\link{latex}}.}

\item{escape}{If \code{TRUE}, then the values of x are
latex escaped.}
}
\value{
An object of class \code{texopts}.
}
\description{
Object representing LaTeX option lists.
}

