% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_latex.R
\docType{data}
\name{URL_REGEX}
\alias{URL_REGEX}
\title{Regex for URL validation}
\format{An object of class \code{character} of length 1.}
\usage{
URL_REGEX
}
\description{
Regular expression for URL validation from Diego Pierri
from this \href{gist}{https://gist.github.com/dperini/729294}.
This is used by the \code{\link{escape_latex}} function to escape URLs.
}
\author{
Diego Pierri. MIT License.
}
\references{
\itemize{
\item \url{https://gist.github.com/dperini/729294}
\item \url{https://mathiasbynens.be/demo/url-regex}
}
}
\seealso{
\code{\link{escape_latex}}
}
\keyword{datasets}

