% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utility-functions}
\alias{braces}
\alias{brackets}
\alias{imath}
\alias{math}
\alias{newline}
\alias{parens}
\alias{texcomment}
\alias{texrow}
\alias{utility-functions}
\alias{verb}
\title{Utility Functions}
\usage{
parens(x)

braces(x)

brackets(x)

math(x, inline = TRUE, dollar = FALSE)

newline(x = character())

imath(x)

texcomment(x, newline = TRUE)

verb(x, delim = c("|", "\\"", "!", "=", "#", "^"))

texrow(x, newline = TRUE)
}
\arguments{
\item{x}{character vector}

\item{inline}{In \code{math}, if \code{TRUE} wraps \code{x} in
display math, e.g., \verb{\[x\]}. Otherwise, the it is returned
as inline math, e.g. \verb{\(x\)}.}

\item{dollar}{In \code{math}, use \verb{$$} for display math,
and \verb{$} for inline math, instead of the defaults.}

\item{newline}{In \code{texcomment}, append a new-line character to each string.}

\item{delim}{Delimiter to use for \code{verb}.
For example, if \code{delim="|"}, then it produce \verb{\\verb|x|}.}
}
\value{
A character vector with the transformed text.
}
\description{
Miscellaneous functions that are useful in latex.
\code{brackets}, \code{braces}, and code \code{paren} wraps
text in square brackets (\verb{[...]}), curly braces (\verb{\{...\}}),
and parentheses (\verb{(...)}), respectively. The \code{math}
function wraps text in LaTeX inline math,
\verb{$...$}, or display math, \verb{\[...\]}.
\code{texcomment} comments out LaTeX text, \verb{\% ...}.
\code{texrow} concates a character string with ampersands,
\verb{x1 & x2 & x3 \\\\}.
}
\details{
\code{newlines} add line breaks (\verb{\\\\}) to each element in the character vector
and concatenates them into a single newline seperated string.
}
\examples{
parens("foo")
braces("foo")
brackets("foo")
math("\\\\frac{1}{2}")
math("\\\\frac{1}{2}", TRUE)
imath("\\\\frac{1}{2}")
math("\\\\frac{1}{2}", dollar = TRUE)
texcomment("commented text")
}

