% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utility-functions}
\alias{braces}
\alias{brackets}
\alias{imath}
\alias{math}
\alias{newlines}
\alias{parens}
\alias{pctcomment}
\alias{utility-functions}
\title{Utility Functions}
\usage{
parens(x)

braces(x)

brackets(x)

math(x, inline = TRUE)

newlines(x)

imath(x)

pctcomment(x, newline = TRUE)
}
\arguments{
\item{x}{character vector}

\item{inline}{In \code{math}, if \code{TRUE} wraps \code{x} in
display math, e.g., \verb{\[x\]}. Otherwise, the it is returned
as inline math, e.g. \verb{$x$}.}

\item{newline}{In \code{comment}, append a new-line character to each string.}
}
\value{
A character vector with the transformed text.
}
\description{
Miscellaneous functions that are useful in latex.
\code{brackets}, \code{braces}, and code \code{paren} wraps
text in square brackets (\verb{[...]}), curly braces (\verb{\{...\}}),
and parentheses (\verb{(...)}), respectively. The \code{math}
function wraps text in LaTeX inline math,
\verb{$...$}, or display math, \verb{\[...\]}.
\code{comment} comments out LaTeX text, \verb{\% ...}.
}
\details{
\code{newlines} add line breaks (\verb{\\\\}) to each element in the character vector
and concatenates them into a single newline seperated string.
}

