% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texcmd.R, R/texutils.R
\name{texmacro}
\alias{brackets}
\alias{dmath}
\alias{group}
\alias{math}
\alias{newline}
\alias{parens}
\alias{texcomment}
\alias{texmacro}
\alias{texrow}
\alias{utility-functions}
\alias{verb}
\title{Utility Functions}
\usage{
texmacro(name, trailing = FALSE, starred = FALSE)

brackets(x, type = "[", size = "auto", ...)

parens(...)

group(x, ...)

math(x, inline = TRUE, dollar = FALSE, ...)

newline(x = character(), ...)

dmath(...)

texcomment(x, newline = TRUE)

verb(x, delim = "|")

texrow(x, newline = TRUE, ...)
}
\arguments{
\item{trailing}{Include trailing curly brackets}

\item{starred}{Make the macro starred}

\item{x}{character vector}

\item{type}{Bracket type}

\item{size}{Auto-size brackets using \verb{\\left} and \verb{\\right}.}

\item{...}{passed to other functions}

\item{inline}{In \code{math}, if \code{TRUE} wraps \code{x} in
display math, e.g., \verb{\[x\]}. Otherwise, the it is returned
as inline math, e.g. \verb{\(x\)}.}

\item{dollar}{In \code{math}, use \verb{$$} for display math,
and \verb{$} for inline math, instead of the defaults.}

\item{newline}{In \code{texcomment}, append a new-line character to each string.}

\item{delim}{Delimiter to use for \code{verb}.
For example, if \code{delim="|"}, then it produce \verb{\\verb|x|}.}
}
\value{
A character vector with the transformed text.

character vector
}
\description{
Miscellaneous functions that are useful in latex.
}
\details{
The function \code{brackets} wraps text in LaTeX brackets.
By default it uses square brackets, \verb{[}, but recognizes
multiple types, and will by default auto-size the brackets with
\verb{\\left} and \verb{\\right}.

The function \code{group} wraps text in curly braces, e.g. \verb{\{x\}},
which is a LaTeX group.

The \code{math} function wraps text in LaTeX inline math,
\verb{$...$}, or display math, \verb{\[...\]}.
The function \code{dmath} is a shortcut for display math, i.e.
\code{math(x, inline=FALSE)}.

\code{texcomment} comments out LaTeX text, \verb{\% ...}.

\code{texrow} concates a character string with ampersands,
\verb{x1 & x2 & x3 \\\\}.

\code{newlines} add line breaks (\verb{\\\\}) to each element in the character vector
and concatenates them into a single newline seperated string.
}
\examples{
parens("foo")
brackets("foo")
brackets("foo", type = "<")
brackets("foo", type = "{")
brackets("foo", size = NULL)
brackets("foo", size = "\\\\bigg")
group("foo")
math("\\\\frac{1}{2}")
math("\\\\frac{1}{2}", TRUE)
imath("\\\\frac{1}{2}")
math("\\\\frac{1}{2}", dollar = TRUE)
texcomment("commented text")
}

